Add support for multiple pipelines to app state (#216)
We are adding the ability to create multiple ('modular') pipelines, and will add a dropdown to allow users to filter by pipeline. This PR adds this information to the application state via the data normalizer, and adds corresponding actions, reducers and selectors to allow the active pipeline to be updated, and the nodes/edges etc to be filtered accordingly.

The UI to allow users to select an active pipeline will be added in a subsequent PR.

I've also added some new tests for the normalizer, as it didn't have any.