Fix visible.layers bug and refactor layer visibility state (#253)
There was an issue where resetting the state (e.g. when async loading data) would override/reset/empty state.visible, as the preparePipelineState function didn't have access to the rest of the state.visible object so it would just replace it rather than merging it. This would cause every other state.visible property to reset to undefined, so for instance the sidebar would close every time the data loaded.

This PR fixes this issue by changing state.visible.layers to state.layer.visible in order to allow this to be handled within a new layer reducer. Additionally, instead of updating state.layer.visible when preparing initial state, instead I'm letting the selectors handle this, and now they will pass an empty array/object if layers are disabled, and the app can treat an empty array of layers as the same as having layers disabled, as they were functionally the same anyway. This allowed me to simplify some of the logic in other places.

I've also:
- refactored the redux store function, to separate out some logic into named functions to make things easier to read.
- removed some unnecessary/irrelevant props from one of the export modal tests.
- deleted the icon-toolbar component. This appears to be left over from a bad merge conflict.