Disable internal CSS in exported SVGs to fix error (#337)
Due to a browser CORS rule, stylesheet cssRules cannot be read from external stylesheets, meaning that the google webfont stylesheet throws an error and breaks when exporting images.

I have an open PR on svg-crowbar to fix this issue, but in the meantime I am temporarily disabling this feature to prevent the error. This will signficantly increase exported SVG filesizes, but it will at least work without throwing an error.