Repeatedly attempt dependency installs on CircleCI (#366)
Running npm ci tends to fail about 5% of the time on Windows in CircleCI. This appears to be an issue with the 'fsevents' subdependency.

Attempts to fix it have failed, so this workaround instructs CircleCI to try repeatedly to install dependencies, and only stop after 5 failed attempts. That way, I figure that if at first it doesn't succeed, it should work on subsequent attempts. If after 5 attempts it still hasn't succeeded, then there is probably a deeper underlying issue and it should throw an error.