Fix plots on metadata panel (#1023)
Resolves #1022

Earlier we were setting self.dataset._version = None if no dataset was found in experiment tracking. If you do that self.dataset.load() function on Kedro will return None when you try to call it from the flowchart. Hence it displayed the error 'is not a directory.'

So i changed this in two places to self.dataset._version = Version(None, None) so now self.dataset.load() will fetch the latest version.