Clean e2e test setup (#1443)
Description
Resolves #805.

Development notes
Ok so the main things done in this PR:-

Added support for the earliest Kedro version that Kedro-viz supports, which is Kedro 0.17.5. The end-to-end (e2e) tests for Kedro 0.17.5 will skip if the Python version is greater than 3.8, as Kedro 18 only works with Python 3.9 and above.
Removed old code related to Kedro 0.16.0 in environment.py, which involved unnecessary package installations.
Implemented a suggestion by @tynandebold to capture Kedro-viz breaking due to old dependencies in our end-to-end (e2e) tests. I created a lower-bound requirements.txt file for testing Kedro-viz functionality. If the tests fail, we know that we need to increase the lower-bound on our dependencies.
Added an additional assert statement to check if kedro-viz is running, as the viz flowchart for Kedro 0.17.5 and Kedro 0.18 are different.
Removed win-e2e-test for Python 3.7 since it was failing. Since we plan to drop support for Python 3.7 soon, we decided not to spend time investigating the issue.
also thanks @astrojuanlu for help on this ticket :)