Inherit tags from nodes for modular pipelines (#1542)
Description
Resolves #795

Development notes
We tried solving this issue in the backend but the way modular pipelines is computed in the backend is very complex. I thought maybe we can try solving it in the FE using selectors.

In this PR, we have a new selector called getTagsforNodesAndModularPipelines.This selector identifies nodes associated with specific tags and then propagates these tags to the corresponding modular pipeline(s) (nested as well) for each node.