Fix `kedro viz --load-file` to run from any directory without requiring a Kedro project (#2206)
This PR fixes #2045.

Previously, kedro viz --load-file checked if the command was run from within a Kedro project directory, which is not required. The fix ensures that kedro viz --load-file can run from anywhere as long as it has the correct path the API data directory .

The key behavior is now as follows:

If the specified filepath does not exist, a ValueError is raised:
ValueError: The provided filepath 'hellod' does not exist.
If the filepath exists but does not contain the expected directory structure (api/main, api/pipelines/, and api/nodes/), a FileNotFoundError is raised:
FileNotFoundError: [Errno 2] No such file or directory: '/path/to/xyz/api/main'

If the filepath exists and has the required api directory structure and conent, kedro viz --load-file runs successfully using the API JSON data.